+++ templates/emails/default/mail.text.php
@@ -1,0 +1,2 @@
+<?= t('philippoehrlein.kirby-email-manager.email.default.message', 'New email from the contact form.') ?>
+<?php snippet('emails/data-text', ['form' => $form]) ?>