+++ templates/emails/default/reply.text.php
@@ -1,0 +1,2 @@
+<?= t('philippoehrlein.kirby-email-manager.email.default.reply_message', 'Thank you for your email.') ?>
+<?php snippet('emails/data-text', ['form' => $form]) ?>